Pass byte-array reference to c-call if possible.
If both the argument and the parameter specify a C-Type byte
array don't copy the data into a new array, instead pass a
pointer directly the called function.

This makes it possible to read data into the passed buffer. This
is needed to make GSocket.receive_from work.

This also means that the buffers of this type can't be released, as
they never were locally copied. The release part will therefore check
and if the parameter is a byte-array and the argument is a byte-array
as well and if so not free the buffer.